ci: deprecate @mono/settings package
