+++ apps/web/package.json
@@ -29,7 +29,6 @@
     "@dotenvx/dotenvx": "^1.10.1",
     "@faker-js/faker": "^8.4.0",
     "@hookform/resolvers": "^3.3.4",
-    "@mono/settings": "workspace:*",
     "@mono/theme": "workspace:*",
     "@mono/tsconfig": "workspace:*",
     "@mono/types": "workspace:*",
