+++ apps/web/src/app/(app)/[locale]/draft/(blog)/blog/[...slug]/page.tsx
@@ -1,5 +1,5 @@
-import type React from 'react';
 import BlogDetail from '@mono/web/app/(app)/[locale]/(blog)/blog/[...slug]/page';
+import type React from 'react';
 
 export const dynamic = 'force-dynamic';
 export const revalidate = 0;
