+++ apps/web/src/app/(app)/[locale]/draft/[[...slug]]/page.tsx
@@ -1,5 +1,5 @@
-import type React from 'react';
 import CatchallPage from '@mono/web/app/(app)/[locale]/[[...slug]]/page';
+import type React from 'react';
 
 export const dynamic = 'force-dynamic';
 export const revalidate = 0;
