+++ apps/web/src/hooks/globalInvalidateCache.ts
@@ -1,4 +1,4 @@
-import { LOCALES } from '@mono/settings';
+import { LOCALES } from '@mono/web/lib/constants';
 import { revalidatePath } from 'next/cache';
 import type { GlobalAfterChangeHook } from 'payload';
