+++ package.json
@@ -9,7 +9,7 @@
     "setup": "turbo run setup",
     "clean": "turbo run clean && rm -rf ./node_modules && rm -rf ./coverage",
     "reinstall": "pnpm clean && pnpm install",
-    "merge-coverage": "mkdir -p coverage && pnpm cobertura-merge -o coverage/coverage.xml web=apps/web/coverage/cobertura-coverage.xml settings=packages/settings/coverage/cobertura-coverage.xml ui=packages/ui/coverage/cobertura-coverage.xml",
+    "merge-coverage": "mkdir -p coverage && pnpm cobertura-merge -o coverage/coverage.xml web=apps/web/coverage/cobertura-coverage.xml ui=packages/ui/coverage/cobertura-coverage.xml",
     "create:block": "npx hygen block new --with-prompt && pnpm generate:types && pnpm lint:fix && cd apps/web && pnpm lint:fix",
     "create:component": "npx hygen component new --with-prompt",
     "create:package": "npx hygen package new --with-prompt",
@@ -23,9 +23,8 @@
     "test:coverage": "vitest run --coverage --coverage.reporter=cobertura",
     "types": "pnpm env:dev -- turbo run types",
     "build": "turbo run build",
-    "buildSettings": "turbo run buildSettings",
     "start": "turbo run start --parallel",
-    "dev": "pnpm env:dev -- turbo run dev --filter=\"!@mono/settings\" --ui=tui",
+    "dev": "pnpm env:dev -- turbo run dev --ui=tui",
     "lint": "pnpm env:dev -- biome check .",
     "lint:fix": "pnpm lint --write",
     "migrations:create": "turbo run migrations:create",
