--- packages/settings/.eslintrc.cjs
@@ -1,8 +1,0 @@
-module.exports = {
-  root: true,
-  extends: ['custom'],
-  ignorePatterns: ['**/dist/**'],
-  rules: {
-    'prefer-destructuring': 'off'
-  }
-};
