--- packages/settings/logSettingsPlugin.ts
@@ -1,16 +1,0 @@
-/* eslint-disable no-console */
-
-// log settings during build
-function logSettingsPlugin() {
-  return {
-    name: 'settings-log',
-    async closeBundle() {
-      // eslint-disable-next-line
-      const SETTINGS = require('./index');
-      console.info('Using the following settings:');
-      console.table(SETTINGS);
-    }
-  };
-}
-
-export default logSettingsPlugin;
