--- packages/settings/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@mono/settings",
-  "version": "0.0.0",
-  "license": "MIT",
-  "type": "module",
-  "main": "./dist/index.umd.cjs",
-  "module": "./dist/index.es.js",
-  "types": "./dist/index.d.ts",
-  "scripts": {
-    "clean": "rm -rf -- ./.turbo ./dist ./coverage ./node_modules",
-    "setup-dev": "tsc && vite build",
-    "build": "tsc && vite build",
-    "buildSettings": "tsc && vite build",
-    "dev": "tsc && vite build --watch",
-    "test": "vitest run",
-    "test:watch": "vitest watch",
-    "test:coverage": "vitest run --coverage",
-    "types": "tsc --noEmit --skipLibCheck"
-  },
-  "exports": {
-    ".": {
-      "import": "./dist/index.es.js",
-      "require": "./dist/index.umd.cjs",
-      "types": "./dist/index.d.ts"
-    }
-  },
-  "dependencies": {
-    "@dotenvx/dotenvx": "^1.10.1",
-    "@mono/tsconfig": "workspace:*",
-    "@vitest/coverage-v8": "latest",
-    "jsdom": "^22.1.0",
-    "typescript": "latest",
-    "unplugin-auto-import": "^0.16.6",
-    "vite": "^5.2.10",
-    "vite-plugin-dts": "latest",
-    "vite-plugin-environment": "latest",
-    "vite-tsconfig-paths": "latest",
-    "vitest": "latest"
-  }
-}
