--- packages/settings/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "@mono/tsconfig/react-library.json",
-  "include": ["."],
-  "exclude": ["dist", "build", "node_modules"],
-  "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true,
-    "types": ["vitest/globals"]
-  }
-}
